Pavel/circular progress - preparation (#1365)
* bootstrap

* move files

* more bootstrapping

* enum rename

* remove donut temporaly

* removed const

* cleanup, added role to side-menu
