+++ projects/ui-framework/src/lib/indicators/progress/progress-bar/progress-bar.component.scss
@@ -1,4 +1,4 @@
-@import '../../style/common-imports';
+@import '../../../style/common-imports';
 @import './progress-bar.mixin';
 
 $BPBtypeMap: (
