+++ projects/ui-framework/src/lib/indicators/simple-bar-chart/simple-bar-chart.component.scss
@@ -1,5 +1,5 @@
 @import '../../style/common-imports';
-@import '../progress-bar/progress-bar.mixin';
+@import '../progress/progress-bar/progress-bar.mixin';
 
 $bsbc-bar-height: 160px;
 $bsbc-bar-width: 40px;
